requirements.txt
pyrogram>=2.0.0
tgcrypto>=1.2.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
sentence-transformers>=2.2.0
numpy<2.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
openai>=1.0.0



